Refactoring Dir Service
