+++ src/Command/CreateFileDefaultCommand.php
@@ -2,7 +2,7 @@
 
 namespace ManageStudent\Command;
 
-use ManageStudent\Service\CommandBanner;
+use ManageStudent\Service\Command\CommandBanner;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
