+++ src/Command/DefaultCommand.php
@@ -3,7 +3,7 @@
 namespace ManageStudent\Command;
 
 use ManageStudent\Service\CkeckStack;
-use ManageStudent\Service\CommandBanner;
+use ManageStudent\Service\Command\CommandBanner;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
