+++ src/Service/FileSystem/FileSelector.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace ManageStudent\Service;
+namespace ManageStudent\Service\FileSystem;
 
+use ManageStudent\Service\chemin;
+use ManageStudent\Service\Command\QuestionType;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\SplFileInfo;
